+++ components/Companies.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { InfiniteMovingCards } from "./ui/infinite-moving-cards";
 
 export default function Companies() {
