+++ package.json
@@ -21,9 +21,9 @@
     "framer-motion": "^12.19.1",
     "lucide-react": "^0.523.0",
     "motion": "^12.19.1",
-    "next": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "next": "^15.3.5",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-icons": "^4.8.0",
     "tailwind-merge": "^3.3.1"
   },
@@ -32,6 +32,7 @@
     "@types/node": "^12.12.21",
     "@types/react": "^18.3.23",
     "@types/react-dom": "^18.3.7",
+    "eslint-config-next": "^15.3.5",
     "postcss-preset-env": "^8.0.1",
     "tailwindcss": "^3.2.7",
     "typescript": "^5.8.3"
